Refactor: Extract DiscussionItem component and improve UI
